Prepare the list with the main collector's last name

Returns the list with the last name of the main collector associated with the unique key recordedBy.

collectors_prepare_dictionary(
  occ = NA,
  collectorDictionary_file =
    "https://raw.githubusercontent.com/pablopains/parseGBIF/main/collectorDictionary/CollectorsDictionary.csv"
)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

collectorDictionary_file

Collector dictionary file - point to a file on your local disk or upload via git at https://raw.githubusercontent.com/pablopains/parseGBIF/main/collectorDictionary/CollectorsDictionary.csv.

Value

Ctrl_nameRecordedBy_Standard, Ctrl_recordedBy, Ctrl_notes, collectorDictionary, Ctrl_update, collectorName, Ctrl_fullName, Ctrl_fullNameII, CVStarrVirtualHerbarium_PersonDetails

Details

If recordedBy is present in the collector's dictionary, it returns the checked name, if not, it returns the last name of the main collector, extracted from the recordedBy field. If recordedBy is present in the collector's dictionary, returns the main collector's last name associated with the single recordedBy key, otherwise, returns the main collector's last name, extracted from the recordedBy field. It is recommended to curate the main collector's surname, automatically extracted from the recordedBy field. The objective is to standardize the last name of the main collector. That the primary botanical collector of a sample is always recognized by the same last name, standardized in capital letters and non-ascii characters replaced

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{
help(collectors_prepare_dictionary)

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)

collectorsDictionaryFromDataset <- collectors_prepare_dictionary(occ = occ,
                                                                collectorDictionary_file =  'https://raw.githubusercontent.com/pablopains/parseGBIF/main/collectorDictionary/CollectorsDictionary.csv')
#> Loading required package: stringr
#> Loading required package: dplyr
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
#> [1] "Loading collectorDictionary..."
#> [1] "Extracting the main collector's surname...."
#> Warning: no non-missing arguments to max; returning -Inf

colnames(collectorsDictionaryFromDataset)
#> [1] "Ctrl_nameRecordedBy_Standard"         
#> [2] "Ctrl_recordedBy"                      
#> [3] "Ctrl_notes"                           
#> [4] "collectorDictionary"                  
#> [5] "Ctrl_update"                          
#> [6] "collectorName"                        
#> [7] "Ctrl_fullName"                        
#> [8] "Ctrl_fullNameII"                      
#> [9] "CVStarrVirtualHerbarium_PersonDetails"
head(collectorsDictionaryFromDataset)
#>   Ctrl_nameRecordedBy_Standard                                 Ctrl_recordedBy
#> 1                      ACEVEDO                         R. ACEVEDO R.; C. REYES
#> 2                       ACOSTA                     R. ACOSTA P.; F. VÁZQUEZ B.
#> 3                       ACOSTA                      R. ACOSTA P.; N. ACOSTA B.
#> 4                       ADARVE                                       J. ADARVE
#> 5                      AGUILAR                                   AGUILAR, R.M.
#> 6                      AGUILAR JOSÉ AGUILAR CANO;SANDRA MEDINA;JHONATAN GUZMÁN
#>   Ctrl_notes collectorDictionary Ctrl_update collectorName Ctrl_fullName
#> 1       <NA>                            <NA>          <NA>          <NA>
#> 2       <NA>                            <NA>          <NA>          <NA>
#> 3       <NA>                            <NA>          <NA>          <NA>
#> 4       <NA>                            <NA>          <NA>          <NA>
#> 5       <NA>                            <NA>          <NA>          <NA>
#> 6       <NA>                            <NA>          <NA>          <NA>
#>   Ctrl_fullNameII CVStarrVirtualHerbarium_PersonDetails
#> 1            <NA>                                  <NA>
#> 2            <NA>                                  <NA>
#> 3            <NA>                                  <NA>
#> 4            <NA>                                  <NA>
#> 5            <NA>                                  <NA>
#> 6            <NA>                                  <NA>

collectorDictionary_checked_file <- paste0(tempdir(),'/','collectorsDictionaryFromDataset.csv')

collectorDictionary_checked_file
#> [1] "C:\\Users\\PABLOH~1\\AppData\\Local\\Temp\\RtmpSSybZ3/collectorsDictionaryFromDataset.csv"

write.csv(collectorsDictionaryFromDataset,
          collectorDictionary_checked_file,
          row.names = FALSE,
          fileEncoding = "UTF-8",
          na = "")
# }