Extracting GBIF issues

Extract GBIF validation rules for occurrence records

GBIF recognises and documents several issues relating to the data fields for an individual record. The issue field stores terms that represent an enumeration of GBIF validation rules. Issues can lead to errors or unexpected data. The issues fields are therefore a valuable source of information when assessing the quality of a record. In order to help GBIF and the data publishers improve the data, GBIF flag records with various issues that they have encountered. These issues can be used as filters applied to occurrence searches. Not all issues indicate bad data, some flagthe fact that GBIF has altered values during processing. The values of EnumOccurrenceIssue will be used by the function extract_gbif_issue as a model to tabulate the GBIF issues of each record, individualizing them, in columns.TRUE or FALSE, flagging whether the issue applies or not for each record.

extract_gbif_issue(occ = NA, enumOccurrenceIssue = NA)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

enumOccurrenceIssue

An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)

Value

list with two data frames: summary, with the frequency of issues in the records and occ_gbif_issue, with issues in columns with TRUE or FALSE for each record.

Details

https://gbif.github.io/parsers/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{

library(ParsGBIF)
#> Error in library(ParsGBIF): there is no package called 'ParsGBIF'

help(extract_gbif_issue)

occ_file <- 'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')
#> Error in open.connection(structure(4L, class = c("curl", "connection"), conn_id = <pointer: 0x0000000000000a2d>),     "rb"): HTTP error 404.

data(EnumOccurrenceIssue)

colnames(EnumOccurrenceIssue)
#> [1] "constant"        "description"     "definition"      "type"           
#> [5] "priority"        "score"           "selection_score" "reasoning"      
#> [9] "notes"          

head(EnumOccurrenceIssue)
#> # A tibble: 6 × 9
#>   constant description definition type  priority score selection_score reasoning
#>   <chr>    <chr>       <chr>      <chr> <chr>    <dbl>           <dbl> <chr>    
#> 1 COORDIN… Indicates … The value… geos… Low          1              -1 Does not…
#> 2 CONTINE… The interp… definitio… geos… Low          1              -1 Does not…
#> 3 CONTINE… The interp… The inter… geos… Low          1              -1 Does not…
#> 4 CONTINE… The interp… If no val… geos… Low          1              -1 Does not…
#> 5 CONTINE… The interp… definitio… geos… Low          1              -1 Does not…
#> 6 COORDIN… Deprecated… definitio… geos… Low          1              -1 Does not…
#> # ℹ 1 more variable: notes <chr>

occ_gbif_issue <- extract_gbif_issue(occ = occ)
#> Error in eval(expr, envir, enclos): object 'occ' not found

names(occ_gbif_issue)
#> Error in eval(expr, envir, enclos): object 'occ_gbif_issue' not found

head(occ_gbif_issue$summary)
#> Error in eval(expr, envir, enclos): object 'occ_gbif_issue' not found

colnames(occ_gbif_issue$occ_gbif_issue)
#> Error in eval(expr, envir, enclos): object 'occ_gbif_issue' not found

head(occ_gbif_issue$occ_gbif_issue)
#> Error in eval(expr, envir, enclos): object 'occ_gbif_issue' not found
# }