Prepare occurrence data from GBIF to use in package

Prepare occurrence data downloaded from GBIF to be used by ParsGBIF functions

prepare_gbif_occurrence_data(gbif_occurrece_file = "", columns = "standard")

Arguments

gbif_occurrece_file

The name of the file from which the with occurrence data downloaded from GBIF (by default "occurrence.txt")

columns

Character vector of strings to indicate column names of the GBIF occurrence file. Use 'standard' to select basic columns for use in the package, 'all' to select all available columns. The default is 'standard'

Value

data.frame with renamed selected fields with prefix "Ctrl_"

Details

Select data fields and rename field names prefixed with "Ctrl_"

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{

library(ParsGBIF)

help(prepare_gbif_occurrence_data)
#> Help on topic 'prepare_gbif_occurrence_data' was found in the following
#> packages:
#> 
#>   Package               Library
#>   ParsGBIF              C:/Users/Pablo
#>                       Hendrigo/AppData/Local/R/win-library/4.3
#>   parseGBIF             C:/Users/Pablo
#>                       Hendrigo/AppData/Local/Temp/RtmpsFfGNV/temp_libpath5178233ad4e
#> 
#> 
#> Using the first match ...

occ_file <- 'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)

colnames(occ)
#>  [1] "Ctrl_bibliographicCitation"          "Ctrl_language"                      
#>  [3] "Ctrl_institutionCode"                "Ctrl_collectionCode"                
#>  [5] "Ctrl_datasetName"                    "Ctrl_basisOfRecord"                 
#>  [7] "Ctrl_informationWithheld"            "Ctrl_dataGeneralizations"           
#>  [9] "Ctrl_occurrenceID"                   "Ctrl_catalogNumber"                 
#> [11] "Ctrl_recordNumber"                   "Ctrl_recordedBy"                    
#> [13] "Ctrl_georeferenceVerificationStatus" "Ctrl_occurrenceStatus"              
#> [15] "Ctrl_eventDate"                      "Ctrl_year"                          
#> [17] "Ctrl_month"                          "Ctrl_day"                           
#> [19] "Ctrl_habitat"                        "Ctrl_fieldNotes"                    
#> [21] "Ctrl_eventRemarks"                   "Ctrl_locationID"                    
#> [23] "Ctrl_higherGeography"                "Ctrl_islandGroup"                   
#> [25] "Ctrl_island"                         "Ctrl_countryCode"                   
#> [27] "Ctrl_stateProvince"                  "Ctrl_county"                        
#> [29] "Ctrl_municipality"                   "Ctrl_locality"                      
#> [31] "Ctrl_verbatimLocality"               "Ctrl_locationRemarks"               
#> [33] "Ctrl_decimalLatitude"                "Ctrl_decimalLongitude"              
#> [35] "Ctrl_verbatimCoordinateSystem"       "Ctrl_verbatimIdentification"        
#> [37] "Ctrl_identificationQualifier"        "Ctrl_typeStatus"                    
#> [39] "Ctrl_identifiedBy"                   "Ctrl_dateIdentified"                
#> [41] "Ctrl_scientificName"                 "Ctrl_family"                        
#> [43] "Ctrl_taxonRank"                      "Ctrl_nomenclaturalCode"             
#> [45] "Ctrl_taxonomicStatus"                "Ctrl_issue"                         
#> [47] "Ctrl_mediaType"                      "Ctrl_hasCoordinate"                 
#> [49] "Ctrl_hasGeospatialIssues"            "Ctrl_verbatimScientificName"        
#> [51] "Ctrl_level0Name"                     "Ctrl_level1Name"                    
#> [53] "Ctrl_level2Name"                     "Ctrl_level3Name"                    

head(occ)
#> # A tibble: 6 × 54
#>   Ctrl_bibliographicCit…¹ Ctrl_language Ctrl_institutionCode Ctrl_collectionCode
#>   <chr>                   <chr>         <chr>                <chr>              
#> 1 NA                      es            Universidad de Anti… HUA                
#> 2 NA                      es            Universidad de Anti… HUA                
#> 3 NA                      es            Universidad de Anti… HUA                
#> 4 NA                      es            Universidad de Anti… HUA                
#> 5 NA                      es            Universidad de Anti… HUA                
#> 6 NA                      es            Universidad de Anti… HUA                
#> # ℹ abbreviated name: ¹​Ctrl_bibliographicCitation
#> # ℹ 50 more variables: Ctrl_datasetName <chr>, Ctrl_basisOfRecord <chr>,
#> #   Ctrl_informationWithheld <chr>, Ctrl_dataGeneralizations <chr>,
#> #   Ctrl_occurrenceID <chr>, Ctrl_catalogNumber <chr>, Ctrl_recordNumber <chr>,
#> #   Ctrl_recordedBy <chr>, Ctrl_georeferenceVerificationStatus <chr>,
#> #   Ctrl_occurrenceStatus <chr>, Ctrl_eventDate <dttm>, Ctrl_year <dbl>,
#> #   Ctrl_month <dbl>, Ctrl_day <dbl>, Ctrl_habitat <chr>, …
# }