Selecting the master digital voucher

To group duplicates and choose the digital voucher: Unique collection events can result in many ‘duplicate’ GBIF records. We designate one of these ‘duplicate’ records as the master digital voucher, to which data from other duplicate vouchers can be merged (see export_data):

Where the collection event key for grouping duplicates is complete, then duplicates can be grouped / parsed. To do so, we evaluate record completeness. Record completeness is calculated based on data-quality scores for the information in the following fields: recordedBy, recordNumber, year, institutionCode, catalogNumber, locality, municipality, countryCode, stateProvince and fieldNotes. The spatial coordinates associated with each duplicate are ranked using a score for the quality of the geospatial information. This score is calculated using the issues listed in the GBIF table, EnumOccurrenceIssue. A score is calculated based on these issues (see above). The duplicate with the highest total score is assigned as the master voucher for the unique collection event. Missing information contained in duplicate records of the unique collection event can then be merged into the master digital voucher (see export_data).

Where the collection event key is incomplete, unique collection event duplicates cannot be parsed. In this case, each record is considered as a unique collection event, without duplicates. However, to know the integrity of the information, record completeness and quality of the geospatial information, are evaluated as described above.

How is the quality score calculated? parseGBIF_digital_voucher = The duplicate with the highest total score, sum of record completeness + quality of geospatial information.

How is record completeness calculated? The quality of the duplicate records associated with each collection event key is measured as the completeness of a record, using the sum of a number of flags (see below) equal to TRUE.

Flags used to calculate record completeness

Is there information about the collector?

Is there information about the collection number?

Is there information about the year of collection?

Is there information about the institution code?

Is there information about the catalog number?

Is there information about the locality?

Is there information about the municipality of collection?

Is there information about the state/province of collection?

Is there information about the field notes?

The quality of geospatial information is based on geographic issues raised by GBIF. GIBF issues relating to geospatial data were classified into three classes based on the data quality scores that we assigned to each of the following GBIF issues recorded in the EnumOccurrenceIssue.

Issue does not affect coordinating accuracy, with selection_score equal to -1

Issue has potential to affect coordinate accuracy, with selection_score equal to -3

Records with a selection_score equal to -9 are excluded.

select_digital_voucher(
  occ = NA,
  occ_gbif_issue = NA,
  occ_wcvp_check_name = NA,
  occ_collectorsDictionary = NA,
  enumOccurrenceIssue = NA,
  silence = TRUE
)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

occ_gbif_issue

= result of function extract_gbif_issue()$occ_gbif_issue

occ_wcvp_check_name

= result of function batch_checkName_wcvp()$occ_wcvp_check_name

occ_collectorsDictionary

= result of function update_collectorsDictionary()$occ_collectorsDictionary

enumOccurrenceIssue

An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)

silence

if TRUE does not display progress messages

Value

list with two data frames: occ_digital voucher_and: occ_digital_voucher, with all data processing fields and occ_results, only result fields.

Details

parseGBIF_duplicates_grouping_status - "groupable", "not groupable: no recordedBy and no recordNumber", "not groupable: no recordNumber" or "not groupable: no recordedBy"

parseGBIF_num_duplicates number of duplicates records

parseGBIF_duplicates TRUE/FALSE

parseGBIF_non_groupable_duplicates TRUE/FALSE

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{
help(select_digital_voucher)

head(occ)
#> Error in eval(expr, envir, enclos): object 'occ' not found
head(res_gbif_issue$occ_gbif_issue)
#> Error in eval(expr, envir, enclos): object 'res_gbif_issue' not found
head(res_checkName_wcvp$occ_wcvp_check_name)
#> Error in eval(expr, envir, enclos): object 'res_checkName_wcvp' not found
head(res_collectorsDictionary$occ_collectorsDictionary)
#> Error in eval(expr, envir, enclos): object 'res_collectorsDictionary' not found
res_digital_voucher_and_sample_identification <- select_digital_voucher(occ = occ,
                                                                        occ_gbif_issue = res_gbif_issue$occ_gbif_issue,
                                                                        occ_wcvp_check_name = res_checkName_wcvp$occ_wcvp_check_name,
                                                                        occ_collectorsDictionary = res_collectorsDictionary$occ_collectorsDictionary,
                                                                        enumOccurrenceIssue = EnumOccurrenceIssue)
#> Loading required package: readr
#> Error in if (is.na(enumOccurrenceIssue)) {    data(EnumOccurrenceIssue)} else {    EnumOccurrenceIssue <- enumOccurrenceIssue}: the condition has length > 1

names(res_digital_voucher_and_sample_identification)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found

head(res_digital_voucher_and_sample_identification$occ_digital_voucher)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found
colnames(res_digital_voucher_and_sample_identification$occ_digital_voucher)
#> Error in eval(expr, envir, enclos): object 'res_digital_voucher_and_sample_identification' not found

# }