standardize_scientificName

standardize binomial name, variety, subspecies, form and hybrids, authorship to allow comparison with names of taxa in the World Checklist of Vascular Plants (WCVP) database

standardize_scientificName(
  searchedName = "Alomia angustata (Gardner) Benth. ex Baker"
)

Arguments

searchedName

scientific name, with or without author

Value

searchedName, standardizeName, taxonAuthors, taxonAuthors_last

Details

Standardize scientific name according to WCVP format. Separate generic epithet, specific epithet, variety, subspecies, form, hybrid and author, in the scientific name, if any. Standardize, according to WCVP, abbreviation of infrataxon, if any: variety to var., subspecies to subsp., FORM to f., hybrid separator separate x from the specific epithet.

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{
# standardize_scientificName()

help(standardize_scientificName)
#> Help on topic 'standardize_scientificName' was found in the following
#> packages:
#> 
#>   Package               Library
#>   ParsGBIF              C:/Users/Pablo
#>                       Hendrigo/AppData/Local/R/win-library/4.3
#>   parseGBIF             C:/Users/Pablo
#>                       Hendrigo/AppData/Local/R/win-library/4.3
#> 
#> 
#> Using the first match ...

standardize_scientificName('Leucanthemum ×superbum (Bergmans ex J.W.Ingram) D.H.Kent')
#> $searchedName
#> [1] "Leucanthemum ×superbum (Bergmans ex J.W.Ingram) D.H.Kent"
#> 
#> $standardizeName
#> [1] "Leucanthemum × superbum"
#> 
#> $taxonAuthors
#> [1] "(Bergmans ex J.W.Ingram) D.H.Kent"
#> 
#> $taxonAuthors_last
#> [1] "D.H.Kent"
#> 
standardize_scientificName('Alomia angustata (Gardner) Benth. ex Baker')
#> $searchedName
#> [1] "Alomia angustata (Gardner) Benth. ex Baker"
#> 
#> $standardizeName
#> [1] "Alomia angustata"
#> 
#> $taxonAuthors
#> [1] "(Gardner) Benth. ex Baker"
#> 
#> $taxonAuthors_last
#> [1] "Benth. ex Baker"
#> 
standardize_scientificName('Centaurea ×aemiliae Font Quer')
#> $searchedName
#> [1] "Centaurea ×aemiliae Font Quer"
#> 
#> $standardizeName
#> [1] "Centaurea × aemiliae"
#> 
#> $taxonAuthors
#> [1] "Font Quer"
#> 
#> $taxonAuthors_last
#> [1] ""
#> 
# }