Use the World Checklist of Vascular Plants (WCVP) database to check accepted names and update synonyms

Use the World Checklist of Vascular Plants database (about WCVP) to check accepted names and update synonyms.

wcvp_check_name(
  searchedName = "Hemistylus brasiliensis Wedd.",
  wcvp_names = "",
  if_author_fails_try_without_combinations = TRUE
)

Arguments

searchedName

scientific name, with or without author

wcvp_names

WCVP table, wcvp_names.csv file from http://sftp.kew.org/pub/data-repositories/WCVP/ If NA, automatically load the latest version of the database by the function parseGBIF::wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names.

if_author_fails_try_without_combinations

option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses

Value

Data frame with WCVP fields

Details

About the World Checklist of Vascular Plants https://powo.science.kew.org/about-wcvp searchNotes values:

Accepted - When only one authorless scientific name is present in the list of TAXON_name with and TAXON_STATUS equal to "Accepted", verified_speciesName = 100.

Accepted among homonyms - When more than one authorless scientific name is present in the TAXON_name list, but only one of the homonyms displays TAXON_STATUS equal to "Accepted", verified_speciesName = number of matches/100.

Homonyms - When more than one authorless scientific name is present in the TAXON_name list and more than one, or none among the homonyms, display TAXON_STATUS equal to "Accepted", verified_speciesName = number of matches/100. Before searching for homonyms, there was a failure in trying to find the matching match between authorless scientific name in TAXON_name and author in TAXON_AUTHORS, in these cases verified_author equal to 0 (zero),

Not Found: When the authorless scientific name is not present in the TAXON_NAME LIST

Unplaced: o When only one authorless scientific name is present in the list of TAXON_name with and TAXON_STATUS = "Unplaced"

Updated: When only one authorless scientific name is present in the list of TAXON_name and ACCEPTED_PLANT_NAME_ID are not empty (and ACCEPTED_PLANT_NAME_ID is different from the ID of the species consulted) taxon_status_of_searchedName, plant_name_id_of_searchedName and taxon_authors_of_searchedName values:

When searchNotes equals "Updated" – The fields record the information of the scientific name originally consulted.

When searchNotes equals "Homonyms" - Fields record the information of homonymous synonyms separated by "|".

verified_author values:

When value equal to 100 – when there is matched match between authorless scientific name in TAXON_name and author in TAXON_AUTHORS.

When value equal to 50 – when there is combined correspondence between authorless scientific name in TAXON_name and author, without (combination), in TAXON_AUTHORS.

When value equal to 0 – regardless of the correspondence between authorless scientific name in TAXON_name, author is not present in TAXON_AUTHORS.

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# These examples take >10 seconds to run and require 'parseGBIF::wcvp_get_data()'
# \donttest{
library(parseGBIF)

help(wcvp_check_name)

wcvp_names <- wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names
#> [1] "downloading: http://sftp.kew.org/pub/data-repositories/WCVP//wcvp.zip"
#> [1] "loading: C:\\Users\\PABLOH~1\\AppData\\Local\\Temp\\RtmpeOzBwT/wcvp_names.csv"
#> [1] " wcvp_names :1419265"
#> Warning: cannot remove file 'C:\Users\PABLOH~1\AppData\Local\Temp\RtmpeOzBwT/downlit', reason 'Permission denied'

# 1) Updated
wcvp_check_name(searchedName = 'Hemistylus brasiliensis Wedd.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 862839            2845408     854167-1         Species          Accepted
#>        wcvp_family wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid
#> 862839  Urticaceae                   Hemistylus                    
#>         wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 862839 boehmerioides                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 862839                                        Benth.                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 862839                Pl. Hartw.                : 123               (1843)
#>        wcvp_nomenclatural_remarks           wcvp_geographic_area
#> 862839                            Colombia to Ecuador, C. Brazil
#>        wcvp_lifeform_description wcvp_climate_description
#> 862839         subshrub or shrub             wet tropical
#>                 wcvp_taxon_name wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 862839 Hemistylus boehmerioides             Benth.                     2845408
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 862839                          NA                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 862839                     NA                   2845407     854167-1
#>        wcvp_hybrid_formula wcvp_reviewed        wcvp_TAXON_NAME_U
#> 862839                                 N HEMISTYLUS BOEHMERIOIDES
#>        wcvp_TAXON_AUTHORS_U             wcvp_searchedName
#> 862839               BENTH. Hemistylus brasiliensis Wedd.
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 862839                           Synonym                            2845409
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 862839                     Wedd. ex Warm.                    0
#>        wcvp_verified_speciesName wcvp_searchNotes
#> 862839                       100          Updated

# 2) Accepted
wcvp_check_name(searchedName = 'Hemistylus boehmerioides Wedd. ex Warm.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 862839            2845408     854167-1         Species          Accepted
#>        wcvp_family wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid
#> 862839  Urticaceae                   Hemistylus                    
#>         wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 862839 boehmerioides                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 862839                                        Benth.                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 862839                Pl. Hartw.                : 123               (1843)
#>        wcvp_nomenclatural_remarks           wcvp_geographic_area
#> 862839                            Colombia to Ecuador, C. Brazil
#>        wcvp_lifeform_description wcvp_climate_description
#> 862839         subshrub or shrub             wet tropical
#>                 wcvp_taxon_name wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 862839 Hemistylus boehmerioides             Benth.                     2845408
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 862839                          NA                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 862839                     NA                   2845407     854167-1
#>        wcvp_hybrid_formula wcvp_reviewed        wcvp_TAXON_NAME_U
#> 862839                                 N HEMISTYLUS BOEHMERIOIDES
#>        wcvp_TAXON_AUTHORS_U                       wcvp_searchedName
#> 862839               BENTH. Hemistylus boehmerioides Wedd. ex Warm.
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 862839                                NA                                 NA
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 862839                                 NA                    0
#>        wcvp_verified_speciesName wcvp_searchNotes
#> 862839                       100         Accepted

# 3) Unplaced - taxon_status = Unplaced
wcvp_check_name(searchedName = 'Leucosyke australis Unruh',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 256961            2497220     854381-1         Species          Unplaced
#>        wcvp_family wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid
#> 256961  Urticaceae                    Leucosyke                    
#>        wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 256961    australis                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 256961                                         Unruh                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 256961         Bot. Jahrb. Syst.              73: 220               (1943)
#>        wcvp_nomenclatural_remarks wcvp_geographic_area
#> 256961                                                
#>        wcvp_lifeform_description wcvp_climate_description     wcvp_taxon_name
#> 256961                      tree                          Leucosyke australis
#>        wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 256961              Unruh                          NA
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 256961                          NA                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 256961                     NA                   2497263     854381-1
#>        wcvp_hybrid_formula wcvp_reviewed   wcvp_TAXON_NAME_U
#> 256961                                 N LEUCOSYKE AUSTRALIS
#>        wcvp_TAXON_AUTHORS_U         wcvp_searchedName
#> 256961                UNRUH Leucosyke australis Unruh
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 256961                                NA                                 NA
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 256961                                 NA                  100
#>        wcvp_verified_speciesName wcvp_searchNotes
#> 256961                       100         Unplaced

# 4) Accepted among homonyms - When author is not informed. In this case, one of the homonyms, taxon_status is accepted
wcvp_check_name(searchedName = 'Parietaria cretica',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 775923            2546201     854904-1         Species          Accepted
#>        wcvp_family wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid
#> 775923  Urticaceae                   Parietaria                    
#>        wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 775923      cretica                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 775923                                            L.                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 775923                   Sp. Pl.               : 1052               (1753)
#>        wcvp_nomenclatural_remarks
#> 775923                           
#>                                                wcvp_geographic_area
#> 775923 Sicilia, Greece to SW. Türkiye, Cyprus, NE. Tunisia to Libya
#>        wcvp_lifeform_description wcvp_climate_description    wcvp_taxon_name
#> 775923       annual or perennial              subtropical Parietaria cretica
#>        wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 775923                 L.                     2546201
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 775923                          NA                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 775923                     NA                   2546177     854904-1
#>        wcvp_hybrid_formula wcvp_reviewed  wcvp_TAXON_NAME_U
#> 775923                                 N PARIETARIA CRETICA
#>        wcvp_TAXON_AUTHORS_U  wcvp_searchedName
#> 775923                   L. Parietaria cretica
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 775923               Misapplied|Accepted                    2546224|2546201
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 775923                           Moris|L.                    0
#>        wcvp_verified_speciesName        wcvp_searchNotes
#> 775923                        50 Accepted among homonyms

# When author is informed
wcvp_check_name(searchedName = 'Parietaria cretica L.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 775923            2546201     854904-1         Species          Accepted
#>        wcvp_family wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid
#> 775923  Urticaceae                   Parietaria                    
#>        wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 775923      cretica                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 775923                                            L.                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 775923                   Sp. Pl.               : 1052               (1753)
#>        wcvp_nomenclatural_remarks
#> 775923                           
#>                                                wcvp_geographic_area
#> 775923 Sicilia, Greece to SW. Türkiye, Cyprus, NE. Tunisia to Libya
#>        wcvp_lifeform_description wcvp_climate_description    wcvp_taxon_name
#> 775923       annual or perennial              subtropical Parietaria cretica
#>        wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 775923                 L.                     2546201
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 775923                          NA                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 775923                     NA                   2546177     854904-1
#>        wcvp_hybrid_formula wcvp_reviewed  wcvp_TAXON_NAME_U
#> 775923                                 N PARIETARIA CRETICA
#>        wcvp_TAXON_AUTHORS_U     wcvp_searchedName
#> 775923                   L. Parietaria cretica L.
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 775923                                NA                                 NA
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 775923                                 NA                  100
#>        wcvp_verified_speciesName wcvp_searchNotes
#> 775923                       100         Accepted

# When author is informed
wcvp_check_name(searchedName = 'Parietaria cretica Moris',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>         wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 1397037            2596260     856545-1         Species          Accepted
#>         wcvp_family wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid
#> 1397037  Urticaceae                   Soleirolia                    
#>         wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 1397037   soleirolii                                          
#>         wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 1397037                      Req.               Dandy                        
#>         wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 1397037            Feddes Repert.                70: 4               (1965)
#>         wcvp_nomenclatural_remarks wcvp_geographic_area
#> 1397037                                      WC. Medit.
#>         wcvp_lifeform_description wcvp_climate_description
#> 1397037                  subshrub              subtropical
#>               wcvp_taxon_name wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 1397037 Soleirolia soleirolii       (Req.) Dandy                     2596260
#>         wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 1397037                     2844953                             
#>         wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 1397037                     NA                   2596264     856545-1
#>         wcvp_hybrid_formula wcvp_reviewed     wcvp_TAXON_NAME_U
#> 1397037                                 N SOLEIROLIA SOLEIROLII
#>         wcvp_TAXON_AUTHORS_U        wcvp_searchedName
#> 1397037          (REQ.)DANDY Parietaria cretica Moris
#>         wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 1397037                        Misapplied                            2546224
#>         wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 1397037                              Moris                  100
#>         wcvp_verified_speciesName wcvp_searchNotes
#> 1397037                       100          Updated

# 5) Homonyms - When author is not informed. In this case, none of the homonyms, taxon_status is Accepted
wcvp_check_name(searchedName = 'Laportea peltata',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>   wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status wcvp_family
#> 1                 NA         <NA>            <NA>              <NA>        <NA>
#>   wcvp_genus_hybrid wcvp_genus wcvp_species_hybrid wcvp_species
#> 1              <NA>       <NA>                <NA>         <NA>
#>   wcvp_infraspecific_rank wcvp_infraspecies wcvp_parenthetical_author
#> 1                    <NA>              <NA>                      <NA>
#>   wcvp_primary_author wcvp_publication_author wcvp_place_of_publication
#> 1                <NA>                    <NA>                      <NA>
#>   wcvp_volume_and_page wcvp_first_published wcvp_nomenclatural_remarks
#> 1                 <NA>                 <NA>                       <NA>
#>   wcvp_geographic_area wcvp_lifeform_description wcvp_climate_description
#> 1                 <NA>                      <NA>                     <NA>
#>   wcvp_taxon_name wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 1            <NA>               <NA>                          NA
#>   wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 1                          NA                         <NA>
#>   wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 1                     NA                        NA         <NA>
#>   wcvp_hybrid_formula wcvp_reviewed wcvp_TAXON_NAME_U wcvp_TAXON_AUTHORS_U
#> 1                <NA>          <NA>              <NA>                 <NA>
#>   wcvp_searchedName wcvp_taxon_status_of_searchedName
#> 1  Laportea peltata                   Synonym|Synonym
#>   wcvp_plant_name_id_of_searchedName wcvp_taxon_authors_of_searchedName
#> 1                    2496933|2496934 Gaudich. & Decne.|(Blume) Gaudich.
#>   wcvp_verified_author wcvp_verified_speciesName wcvp_searchNotes
#> 1                    0                        50         Homonyms

# When author is informed
wcvp_check_name(searchedName = 'Laportea peltata Gaudich. & Decne.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 342925            2760699     851455-1         Species          Accepted
#>        wcvp_family wcvp_genus_hybrid  wcvp_genus wcvp_species_hybrid
#> 342925  Urticaceae                   Dendrocnide                    
#>        wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 342925   morobensis                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 342925                                          Chew                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 342925     Gard. Bull. Singapore               25: 67               (1969)
#>        wcvp_nomenclatural_remarks wcvp_geographic_area
#> 342925                                  NE. New Guinea
#>        wcvp_lifeform_description wcvp_climate_description
#> 342925                      tree             wet tropical
#>               wcvp_taxon_name wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 342925 Dendrocnide morobensis               Chew                     2760699
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 342925                          NA                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 342925                     NA                   2760664     851455-1
#>        wcvp_hybrid_formula wcvp_reviewed      wcvp_TAXON_NAME_U
#> 342925                                 N DENDROCNIDE MOROBENSIS
#>        wcvp_TAXON_AUTHORS_U                  wcvp_searchedName
#> 342925                 CHEW Laportea peltata Gaudich. & Decne.
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 342925                           Synonym                            2496933
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 342925                  Gaudich. & Decne.                  100
#>        wcvp_verified_speciesName wcvp_searchNotes
#> 342925                       100          Updated

# When author is informed
wcvp_check_name(searchedName = 'Laportea peltata (Blume) Gaudich.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
#>        wcvp_plant_name_id wcvp_ipni_id wcvp_taxon_rank wcvp_taxon_status
#> 342934            2760704     851460-1         Species          Accepted
#>        wcvp_family wcvp_genus_hybrid  wcvp_genus wcvp_species_hybrid
#> 342934  Urticaceae                   Dendrocnide                    
#>        wcvp_species wcvp_infraspecific_rank wcvp_infraspecies
#> 342934      peltata                                          
#>        wcvp_parenthetical_author wcvp_primary_author wcvp_publication_author
#> 342934                     Blume                Miq.                        
#>        wcvp_place_of_publication wcvp_volume_and_page wcvp_first_published
#> 342934                Pl. Jungh.                 : 30               (1851)
#>        wcvp_nomenclatural_remarks   wcvp_geographic_area
#> 342934                            S. Malesia, New Guinea
#>        wcvp_lifeform_description wcvp_climate_description     wcvp_taxon_name
#> 342934                      tree             wet tropical Dendrocnide peltata
#>        wcvp_taxon_authors wcvp_accepted_plant_name_id
#> 342934       (Blume) Miq.                     2760704
#>        wcvp_basionym_plant_name_id wcvp_replaced_synonym_author
#> 342934                     2448925                             
#>        wcvp_homotypic_synonym wcvp_parent_plant_name_id wcvp_powo_id
#> 342934                     NA                   2760664     851460-1
#>        wcvp_hybrid_formula wcvp_reviewed   wcvp_TAXON_NAME_U
#> 342934                                 N DENDROCNIDE PELTATA
#>        wcvp_TAXON_AUTHORS_U                 wcvp_searchedName
#> 342934          (BLUME)MIQ. Laportea peltata (Blume) Gaudich.
#>        wcvp_taxon_status_of_searchedName wcvp_plant_name_id_of_searchedName
#> 342934                           Synonym                            2496934
#>        wcvp_taxon_authors_of_searchedName wcvp_verified_author
#> 342934                   (Blume) Gaudich.                  100
#>        wcvp_verified_speciesName wcvp_searchNotes
#> 342934                       100          Updated
# }