In batch, use the WCVP database to check accepted names and update synonyms

Species’ names can be checked against WCVP database one by one, or in a batch mode. To verify individual names, the function wcvp_check_name is used.

wcvp_check_name_batch(
  occ = NA,
  wcvp_names = "",
  if_author_fails_try_without_combinations = TRUE,
  wcvp_selected_fields = "standard",
  silence = TRUE
)

Arguments

occ

GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')

wcvp_names

get data frame in parseGBIF::wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names or configure function to save a copy on local disk to optimize loading, see details in help(wcvp_get_data)

if_author_fails_try_without_combinations

option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses.

wcvp_selected_fields

WCVP fields selected as return, 'standard' basic columns, 'all' all available columns. The default is 'standard'

silence

if TRUE does not display progress messages

Value

list with two data frames: summary, species list and occ_wcvp_check_name, with WCVP fields

Details

See help(checkName_wcvp)

about WCVP database

World Checklist of Vascular Plants

WCVP database

(about WCVP)

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# These examples take >10 minutes to run and require 'parseGBIF::wcvp_get_data()'
# \donttest{

library(parseGBIF)

help(wcvp_check_name_batch)

occ_file <- 'https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)

# wcvp_names <- wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names
data(wcvp_names_Achatocarpaceae)

head(wcvp_names)
#> Error in eval(expr, envir, enclos): object 'wcvp_names' not found

res_wcvp_check_name_batch <- wcvp_check_name_batch(occ = occ,
                                                 wcvp_names =  wcvp_names,
                                                 if_author_fails_try_without_combinations = TRUE,
                                                 wcvp_selected_fields = 'standard',
                                                 show_process = TRUE)
#> Error in wcvp_check_name_batch(occ = occ, wcvp_names = wcvp_names, if_author_fails_try_without_combinations = TRUE,     wcvp_selected_fields = "standard", show_process = TRUE): unused argument (show_process = TRUE)

names(res_wcvp_check_name_batch)
#> Error in eval(expr, envir, enclos): object 'res_wcvp_check_name_batch' not found

head(res_wcvp_check_name_batch$summary)
#> Error in eval(expr, envir, enclos): object 'res_wcvp_check_name_batch' not found

head(res_wcvp_check_name_batch$occ_wcvp_check_name)
#> Error in eval(expr, envir, enclos): object 'res_wcvp_check_name_batch' not found

# }