Get WCVP database

Download World Checklist of Vascular Plants (WCVP) database

wcvp_get_data(
  url_source = "http://sftp.kew.org/pub/data-repositories/WCVP/",
  read_only_to_memory = FALSE,
  path_results = "C:/parseGBIF",
  update = FALSE,
  load_distribution = FALSE
)

Arguments

url_source

http://sftp.kew.org/pub/data-repositories/WCVP/

read_only_to_memory

TRUE to in-memory read-only, not writing a copy to local disk

path_results

download destination folder, if read_only_to_memory FALSE

update

TRUE to update and load files, FALSE to keep local version and load files, if read_only_to_memory FALSE

load_distribution

TRUE to load file with geographical distribution of species, if read_only_to_memory FALSE

Value

list with two data frames: wcvp_names and wcvp_distribution

Details

http://sftp.kew.org/pub/data-repositories/WCVP/ This is the public SFTP (Secure File Transfer Protocol) site of the Royal Botanic Gardens, Kew. This space contains data resources publicly accessible to the user `anonymous'. No password required for access. Use of data made available via this site may be subject to legal and licensing restrictions. The README in the top-level directory for each data resource provides specific information about its terms of use.

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{
# load package
library(parseGBIF)

help(wcvp_get_data)

# Download wcvp database to local disk
path_data <- tempdir() # you can change this folder

wcvp <- wcvp_get_data(url_source = 'http://sftp.kew.org/pub/data-repositories/WCVP/',
                 read_only_to_memory = FALSE,
                 path_results = path_data,
                 update = FALSE,
                 load_distribution = TRUE)
#> [1] "downloading: http://sftp.kew.org/pub/data-repositories/WCVP//wcvp.zip"
#> [1] "loading: C:\\Users\\PABLOH~1\\AppData\\Local\\Temp\\RtmpSSybZ3/dataWCVP/wcvp_names.csv"
#> [1] " wcvp_names :1419265"
#> [1] "loading: C:\\Users\\PABLOH~1\\AppData\\Local\\Temp\\RtmpSSybZ3/dataWCVP/wcvp_distribution.csv"
#> [1] "wcvp_distribution :1956333"

names(wcvp)
#> [1] "wcvp_names"        "wcvp_distribution"

head(wcvp$wcvp_names)
#>   plant_name_id  ipni_id taxon_rank taxon_status    family genus_hybrid
#> 1        195508 243233-2    Species      Synonym Lamiaceae             
#> 2        197585 767122-1    Species      Synonym Rubiaceae             
#> 3         76791 595920-1    Species      Synonym Myrtaceae             
#> 4         74373 593644-1    Species      Synonym Myrtaceae             
#> 5        205204 884387-1 Subspecies      Synonym Lamiaceae             
#> 6        102745 326216-2      Genus      Synonym Lamiaceae             
#>         genus species_hybrid     species infraspecific_rank infraspecies
#> 1     Stachys                  pustulosa                                
#> 2 Stenostomum                 dichotomum                                
#> 3     Eugenia                   scoparia                                
#> 4     Eugenia                   areolata                                
#> 5      Thymus                pallasianus             subsp.   brachyodon
#> 6    Isanthus                                                           
#>   parenthetical_author primary_author publication_author place_of_publication
#> 1                               Rydb.                               Brittonia
#> 2                                 DC.                                  Prodr.
#> 3                              Duthie         J.D.Hooker      Fl. Brit. India
#> 4                  DC.         Duthie         J.D.Hooker      Fl. Brit. India
#> 5               Borbás          Jalas                      Bot. J. Linn. Soc.
#> 6                              Michx.                          Fl. Bor.-Amer.
#>   volume_and_page first_published nomenclatural_remarks geographic_area
#> 1           1: 95          (1931)                                      
#> 2          4: 461          (1830)                                      
#> 3          2: 489          (1878)                                      
#> 4          2: 490          (1878)                                      
#> 5         64: 262          (1971)                                      
#> 6            2: 3          (1803)                                      
#>   lifeform_description climate_description                           taxon_name
#> 1                                                             Stachys pustulosa
#> 2                                                        Stenostomum dichotomum
#> 3                                                              Eugenia scoparia
#> 4                                                              Eugenia areolata
#> 5                                          Thymus pallasianus subsp. brachyodon
#> 6                                                                      Isanthus
#>    taxon_authors accepted_plant_name_id basionym_plant_name_id
#> 1          Rydb.                 195467                     NA
#> 2            DC.                 197582                     NA
#> 3         Duthie                 199254                     NA
#> 4   (DC.) Duthie                 200472                 199231
#> 5 (Borbás) Jalas                 204938                 204517
#> 6         Michx.                 208472                     NA
#>   replaced_synonym_author homotypic_synonym parent_plant_name_id  powo_id
#> 1                                        NA                   NA 243233-2
#> 2                                        NA                   NA 767122-1
#> 3                                        NA                   NA 595920-1
#> 4                                        NA                   NA 593644-1
#> 5                                        NA                   NA 884387-1
#> 6                                        NA                   NA 326216-2
#>   hybrid_formula reviewed                         TAXON_NAME_U TAXON_AUTHORS_U
#> 1                       Y                    STACHYS PUSTULOSA           RYDB.
#> 2                       Y               STENOSTOMUM DICHOTOMUM             DC.
#> 3                       Y                     EUGENIA SCOPARIA          DUTHIE
#> 4                       Y                     EUGENIA AREOLATA     (DC.)DUTHIE
#> 5                       Y THYMUS PALLASIANUS SUBSP. BRACHYODON   (BORBÁS)JALAS
#> 6                       Y                             ISANTHUS          MICHX.
colnames(wcvp$wcvp_names)
#>  [1] "plant_name_id"           "ipni_id"                
#>  [3] "taxon_rank"              "taxon_status"           
#>  [5] "family"                  "genus_hybrid"           
#>  [7] "genus"                   "species_hybrid"         
#>  [9] "species"                 "infraspecific_rank"     
#> [11] "infraspecies"            "parenthetical_author"   
#> [13] "primary_author"          "publication_author"     
#> [15] "place_of_publication"    "volume_and_page"        
#> [17] "first_published"         "nomenclatural_remarks"  
#> [19] "geographic_area"         "lifeform_description"   
#> [21] "climate_description"     "taxon_name"             
#> [23] "taxon_authors"           "accepted_plant_name_id" 
#> [25] "basionym_plant_name_id"  "replaced_synonym_author"
#> [27] "homotypic_synonym"       "parent_plant_name_id"   
#> [29] "powo_id"                 "hybrid_formula"         
#> [31] "reviewed"                "TAXON_NAME_U"           
#> [33] "TAXON_AUTHORS_U"        

head(wcvp$wcvp_distribution)
#>   plant_locality_id plant_name_id continent_code_l1        continent
#> 1           1394111             1                 8 SOUTHERN AMERICA
#> 2           1394112             1                 8 SOUTHERN AMERICA
#> 3           1394113             1                 8 SOUTHERN AMERICA
#> 4           1394114             1                 8 SOUTHERN AMERICA
#> 5           1394115             1                 8 SOUTHERN AMERICA
#> 6           1394116             1                 8 SOUTHERN AMERICA
#>   region_code_l2                 region area_code_l3       area introduced
#> 1             80        Central America          COS Costa Rica          0
#> 2             82 Northern South America          VEN  Venezuela          0
#> 3             83  Western South America          BOL    Bolivia          0
#> 4             83  Western South America          CLM   Colombia          0
#> 5             83  Western South America          ECU    Ecuador          0
#> 6             83  Western South America          PER       Peru          0
#>   extinct location_doubtful
#> 1       0                 0
#> 2       0                 0
#> 3       0                 0
#> 4       0                 0
#> 5       0                 0
#> 6       0                 0
colnames(wcvp$wcvp_distribution)
#>  [1] "plant_locality_id" "plant_name_id"     "continent_code_l1"
#>  [4] "continent"         "region_code_l2"    "region"           
#>  [7] "area_code_l3"      "area"              "introduced"       
#> [10] "extinct"           "location_doubtful"
# }