Export of results

For each unique collection event key, complete or incomplete, outputs will be created which combine information from duplicate records and generate a single unique collection event record to replace them. The main output fields relating to taxonomic identification and geographic coordinates:

parseGBIF_sample_taxon_name = scientific name chosen as taxonomic identification for unique collection event

parseGBIF_number_taxon_names = number of scientific names found in duplicates of unique collection event

parseGBIF_sample_taxon_name_status = status of choice of 'identified', 'divergent identifications', 'unidentified'

parseGBIF_unidentified_sample = if unique collection event has taxonomic identification

parseGBIF_decimalLatitude = latitude in decimal degrees

parseGBIF_decimalLongitude = longitude in decimal degrees

parseGBIF_useful_for_spatial_analysis = whether the coordinates are useful for spatial analysis. How is the taxon binomial attributed to the unique collection event selected?

Where the unique collection event key is complete: The accepted TAXON_NAME selected is that which is most frequently applied to the duplicate vouchers at or below the rank of species. Where two named are applied with equal frequency then a mechanical approach, using alphabetical order, is applied, the first listed TAXON_NAME being chosen. Where there is no identification, at or below the rank of species, then the unique collection event, the unique collection event is indicated as unidentified.

Where the unique collection event key is incomplete: Where the unique collection event key is incomplete, then each record is treated as a unique collection event. If there is no identification, at or below the rank of species, then the unique collection event is classified as unidentified. __Geospatial information __ If the master voucher does not have geographic coordinates, we will seek coordinates from the duplicate records associated with it. Finally, the records are separated into three sets of data: useable_data - Where unique collection event with taxonomic identification and geographic coordinates are complete. This represents the useable dataset. unusable_data - Where unique collection event without taxonomic identification and/or geographic coordinates. duplicates The duplicates of unique collection events complete / incomplete

With this, it is possible to perform: Merge information between fields of duplicates of a unique collection event to create a synthetic record for each unique collection event, Compare the frequency of content in fields Generate a work package summary

For each complete unique collection event key, data fields that are empty in the digital voucher record will be populated with data from the respective duplicates. During content merging, we indicate fields associated with the description, location, and data of the unique collection event. By default, fields_to_merge parameter of export_data function contains:

Ctrl_fieldNotes

Ctrl_year

Ctrl_stateProvince

Ctrl_municipality

Ctrl_locality

Ctrl_countryCode

Ctrl_eventDate

Ctrl_habitat

Ctrl_level0Name

Ctrl_level1Name

Ctrl_level2Name

Ctrl_level3Name

export_data(
  occ_digital_voucher_file = "",
  occ_digital_voucher = NA,
  merge_unusable_data = FALSE,
  fields_to_merge = c("Ctrl_fieldNotes", "Ctrl_year", "Ctrl_stateProvince",
    "Ctrl_municipality", "Ctrl_locality", "Ctrl_countryCode", "Ctrl_eventDate",
    "Ctrl_habitat", "Ctrl_level0Name", "Ctrl_level1Name", "Ctrl_level2Name",
    "Ctrl_level3Name"),
  fields_to_compare = c("Ctrl_gbifID", "Ctrl_scientificName", "Ctrl_recordedBy",
    "Ctrl_recordNumber", "Ctrl_identifiedBy", "Ctrl_dateIdentified",
    "Ctrl_institutionCode", "Ctrl_collectionCode", "Ctrl_datasetName",
    "Ctrl_datasetName", "Ctrl_language", "wcvp_plant_name_id", "wcvp_taxon_rank",
    "wcvp_taxon_status", "wcvp_family", "wcvp_taxon_name", "wcvp_taxon_authors",
    "wcvp_reviewed", "wcvp_searchNotes"),
  fields_to_parse = c("Ctrl_gbifID", "Ctrl_bibliographicCitation", "Ctrl_language",
    "Ctrl_institutionCode", "Ctrl_collectionCode", "Ctrl_datasetName",
    "Ctrl_basisOfRecord", "Ctrl_catalogNumber", "Ctrl_recordNumber", "Ctrl_recordedBy",
    "Ctrl_occurrenceStatus", "Ctrl_eventDate", "Ctrl_year", "Ctrl_month", "Ctrl_day",
    "Ctrl_habitat", "Ctrl_fieldNotes", "Ctrl_eventRemarks", "Ctrl_countryCode",
    "Ctrl_stateProvince", "Ctrl_municipality", "Ctrl_county", "Ctrl_locality",
    "Ctrl_level0Name", "Ctrl_level1Name", "Ctrl_level2Name", 
     "Ctrl_level3Name",
    "Ctrl_identifiedBy", "Ctrl_dateIdentified", "Ctrl_scientificName",
    "Ctrl_decimalLatitude", "Ctrl_decimalLongitude", "Ctrl_nameRecordedBy_Standard",
    "Ctrl_recordNumber_Standard", "Ctrl_key_family_recordedBy_recordNumber",
    "Ctrl_geospatial_quality", "Ctrl_verbatim_quality", "Ctrl_moreInformativeRecord",
    "Ctrl_coordinates_validated_by_gbif_issue", "wcvp_plant_name_id", "wcvp_taxon_rank",
    "wcvp_taxon_status", "wcvp_family", "wcvp_taxon_name", "wcvp_taxon_authors",
    "wcvp_reviewed", 
     "wcvp_searchNotes", "parseGBIF_digital_voucher",
    "parseGBIF_duplicates", "parseGBIF_num_duplicates",
    "parseGBIF_non_groupable_duplicates", "parseGBIF_duplicates_grouping_status"),
  silence = TRUE
)

Arguments

occ_digital_voucher_file

CSV fila result of function select_digital_voucher()$occ_digital_voucher

occ_digital_voucher

data frame result of function select_digital_voucher()$occ_digital_voucher

merge_unusable_data

include records unique collection events incomplete in merge processing

fields_to_merge

fields to merge

fields_to_compare

fields to compare content frequency

fields_to_parse

all fields

silence

if TRUE does not display progress messages

Value

list with 10 data frames

all_data All records processed, merged Unique collection events complete / incomplete and their duplicates

useable_data_merge Merged Unique collection events complete

useable_data_raw Raw Unique collection events complete

duplicates Duplicates of unique collection events complete / incomplete

unusable_data_merge Merged Unique collection events incomplete, It is NA if merge_unusable_data is FALSE.

unusable_data_raw Raw Unique collection events incomplete

parseGBIF_general_summary

parseGBIF_merge_fields_summary

parseGBIF_merge_fields_summary_useable_data

parseGBIF_merge_fields_summary_unusable_data It is NA if merge_unusable_data is FALSE

Details

Each data frame should be used as needed

See also

Author

Pablo Hendrigo Alves de Melo, Nadia Bystriakova & Alexandre Monro

Examples

# \donttest{
help(export_data)

results <- export_data(occ_digital_voucher_file = file.occ_digital_voucher,
                       merge_unusable_data = TRUE)
#> Error in eval(expr, envir, enclos): object 'file.occ_digital_voucher' not found

names(results)
#> Error in eval(expr, envir, enclos): object 'results' not found

results$parseGBIF_general_summary
#> Error in eval(expr, envir, enclos): object 'results' not found
results$parseGBIF_merge_fields_summary
#> Error in eval(expr, envir, enclos): object 'results' not found
results$parseGBIF_merge_fields_summary_complete
#> Error in eval(expr, envir, enclos): object 'results' not found
NROW(results$all_data)
#> Error in eval(expr, envir, enclos): object 'results' not found
NROW(results$unique_collection_event_complete_merge)
#> Error in eval(expr, envir, enclos): object 'results' not found
NROW(results$unique_collection_event_incomplete_raw)
#> Error in eval(expr, envir, enclos): object 'results' not found
NROW(results$duplicates)
#> Error in eval(expr, envir, enclos): object 'results' not found

# }